field_pipeline/requirements.txt
typer[all]
pyyaml
geopandas
shapely
fiona
rasterio
duckdb
pytest
openai

